requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
python-dotenv
openai
streamlit
schedule
asyncpg
python-multipart
aiofiles
Pillow
streamlit-extras

pytest
pytest-asyncio
respx>=0.21
starlette
anyio
SQLAlchemy>=2
aiosqlite